Clean up after rebase conflicts
